moved dashboards to seperate py files
