+++ front_end/main.py
@@ -18,6 +18,16 @@ import plotly.express as px
 # import os
 import os
 
+# import sys
+import sys
+
+# import dashboards from local files
+import front_end.dashboards.residential_dashboard as residential_dashboard
+import front_end.dashboards.service_dashboard as service_dashboard
+import front_end.dashboards.structure_dashboard as structure_dashboard
+import front_end.dashboards.industrial_dashboard as industrial_dashboard
+import front_end.dashboards.facade_dashboard as facade_dashboard
+
 #--------------------------
 #PAGE CONFIG
 st.set_page_config(
@@ -443,706 +453,18 @@ if show_team_specific_metrics:
 
     # Display team-specific metrics based on selection
     if selected_team == "Residential":
-        metric_col1.metric("Unit Types", "4")
-        metric_col2.metric("Normalized Travel Time", "0.8 (hours)")
-
-        number_of_units = [40, 60, 30, 20]  
-        unit_types = ['Housing', 'Social', 'Commercial', 'Open Space']
-        total_number_of_units = sum(number_of_units)
-
-        # Compute the formula
-        numerator = sum(
-            units * (units - 1) for units in number_of_units
-        )
-        denominator = total_number_of_units * (total_number_of_units - 1)
-
-        if denominator != 0:  # Avoid division by zero
-            calculated_value = 1 - (numerator / denominator)
-        else:
-            calculated_value = None  # Handle division by zero case
-
-        # Residential-specific chart
-        residential_data = pd.DataFrame({
-            'Unit Type': unit_types,
-            'Number of Units': number_of_units
-        })
-        res_chart = px.bar(residential_data, x='Unit Type', y='Number of Units')
-        res_chart.update_layout(
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-        st.plotly_chart(res_chart, use_container_width=True)
-
-        st.markdown("""
-        <h2 style='text-align: center, font-size: 24px;'>Mixed Use Index</h2>
-        """, unsafe_allow_html=True)
-
-        st.markdown("")
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the metric can be expressed as:
-            
-            $$ 
-            1 - \frac{\sum \text{NumberOfUnitsOfASingleFunction}_i \cdot (\text{NumberOfUnitsOfASingleFunction}_i - 1)}{\text{TotalNumberOfUnits} \cdot (\text{TotalNumberOfUnits} - 1)}
-            $$
-        """, unsafe_allow_html=True)
-
-        st.markdown("")
-
-        # Display the computed value
-        if calculated_value is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Value: {calculated_value:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
+        residential_dashboard.run(metric_col1, metric_col2, selected_team)
 
     elif selected_team == "Service":
-        occupancy_efficiency = 80
-        utilization_rate = 34
-        n = 5
-        active_hours = 12
-        function_exchange_factor = 4
-        total_available_hours_per_day = 13
-        total_spaces_available = 50
-
-        metric_col1.metric("Number of Spaces", n)
-
-        metrics_data = {
-            "Occupancy Efficiency": occupancy_efficiency,  # No units here for calculations
-            "Utilization Rate": utilization_rate,  # No units here for calculations
-            "Active Hours": active_hours,  # No units here for calculations
-            "Function Exchange Factor": function_exchange_factor,  # No units here for calculations
-            "Total Available Hours per Day": total_available_hours_per_day,  # No units here for calculations
-            "Total Spaces Available": total_spaces_available  # No units here for calculations
-        }
-
-        # Create columns for metrics based on the number of metrics to display
-        metric_cols = st.columns(len(metrics_data))  # Create a column for each metric
-
-        for i, (metric_name, value) in enumerate(metrics_data.items()):
-            col = metric_cols[i]  # Access the corresponding column
-
-            # Display the metric with units in the name
-            if metric_name == "Occupancy Efficiency":
-                col.metric(metric_name, f"{value}%", help="Percentage of occupancy efficiency")
-            elif metric_name == "Utilization Rate":
-                col.metric(metric_name, f"{value}%", help="Percentage of utilization rate")
-            elif metric_name == "Active Hours":
-                col.metric(metric_name, f"{value} hours", help="Total active hours")
-            elif metric_name == "Function Exchange Factor":
-                col.metric(metric_name, f"{value}", help="Unitless function exchange factor")
-            elif metric_name == "Total Available Hours per Day":
-                col.metric(metric_name, f"{value} hours", help="Total available hours per day")
-            elif metric_name == "Total Spaces Available":
-                col.metric(metric_name, f"{value}", help="Total number of spaces available")
-
-            # Ensure value is a valid number before conversion
-            try:
-                # Use the raw value for calculations
-                raw_value = float(value)  # Convert to float for calculations
-                if metric_name in ["Occupancy Efficiency", "Utilization Rate"]:
-                    max_value = 100
-                elif metric_name == "Active Hours":
-                    max_value = 24
-                elif metric_name == "Function Exchange Factor":
-                    max_value = 10
-                elif metric_name == "Total Available Hours per Day":
-                    max_value = 24
-                elif metric_name == "Total Spaces Available":
-                    max_value = 100
-                
-                progress_bar = col.progress(raw_value / max_value)  # Use raw_value for progress bar
-            except (ValueError, TypeError) as e:
-                print(f"Error converting value for {metric_name}: {value}. Error: {e}")
-                progress_bar = col.progress(0)  # Default to 0 if conversion fails
-
-            col.markdown(f"""
-            <div style="text-align: center; font-size: 24px;">
-            {value} / {max_value}
-            </div>
-            """, unsafe_allow_html=True)
-
-        st.markdown("")
-
-
-        # Compute the formula
-        numerator = utilization_rate * active_hours * function_exchange_factor # what kind of summation we need to do here
-        denominator = total_available_hours_per_day * total_spaces_available
-
-        if denominator != 0:  # Avoid division by zero
-            calculated_value = numerator / denominator
-        else:
-            calculated_value = None  # Handle division by zero case
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the metric can be expressed as:
-            
-            $$ 
-            \frac{\sum_{i=1}^{n} (UtilizationRateOfFunction_i \cdot ActiveHoursOfFunctionPerDay_i \cdot FunctionExchangeFactor)}{TotalAvailableHoursPerDay \cdot TotalSpacesAvailable}
-            $$
-        """, unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-        if calculated_value is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Occupancy Efficiency: {calculated_value:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-
+        service_dashboard.run(metric_col1, metric_col2, selected_team)
+        
     elif selected_team == "Structure":
-
-        #floor flexibility:column-free FAR
-        # Define the variables for the new formula
-        total_column_free_floor_area = 800  # Example value for total column-free floor area
-        total_floor_area = 1000  # Example value for total floor area
-
-        # Compute the formula
-        if total_floor_area != 0:  # Avoid division by zero
-            column_free_floor_area_ratio = total_column_free_floor_area / total_floor_area
-        else:
-            column_free_floor_area_ratio = None  # Handle division by zero case
-
-        # Structure-specific chart
-
-        st.markdown("<h2 style='text-align: center;'>Floor Flexibility: Column-Free FAR</h2>", unsafe_allow_html=True)
-        structure_data = pd.DataFrame({
-            'Number of Columns': [10],
-            'Total Floor Area': [1000]
-        })
-        structure_chart = px.bar(structure_data, x='Number of Columns', y='Total Floor Area')
-        structure_chart.update_layout(
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-        st.plotly_chart(structure_chart, use_container_width=True)
-
-        # Add major text underneath the metrics
-        st.markdown("<h2 style='text-align: center; font-size: 24px;'>Floor Flexibility: Column-Free Floor Area Ratio</h2>", unsafe_allow_html=True)
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the Column-Free Floor Area Ratio can be expressed as:
-            
-            $$ 
-            \text{Column-Free Floor Area Ratio} = \frac{\text{Total Column-Free Floor Area (m²)}}{\text{Total Floor Area (m²)}}
-            $$
-        """, unsafe_allow_html=True)
-
-        # Display the computed value
-        if column_free_floor_area_ratio is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Value: {column_free_floor_area_ratio:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        #structural efficiency: embodied carbon emissions per square meter
-        # Define the variables for the new formula
-        total_embodied_carbon_emissions = 800  
-        usable_floor_area = 1000  
-
-        # Compute the formula
-        if usable_floor_area != 0:  # Avoid division by zero
-            embodied_carbon_emissions_per_square_meter = total_embodied_carbon_emissions / usable_floor_area
-        else:
-            embodied_carbon_emissions_per_square_meter = None  # Handle division by zero case
-
-        # Structure-specific chart
-
-        st.markdown("<h2 style='text-align: center;'>Structural Efficiency: Embodied Carbon Emissions per Square Meter</h2>", unsafe_allow_html=True)
-        structure_data = pd.DataFrame({
-            'Total Embodied Carbon Emmissions': [10],
-            'Floor Area': [1000]
-        })
-        structure_chart = px.bar(structure_data, x='Total Embodied Carbon Emmissions', y='Floor Area')
-        structure_chart.update_layout(
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-        st.plotly_chart(structure_chart, use_container_width=True)
-
-        # Add major text underneath the metrics
-        st.markdown("<h2 style='text-align: center; font-size: 24px;'>Structural Flexibility: Embodied Carbon Emissions per Square Meter</h2>", unsafe_allow_html=True)
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the Embodied Carbon Emissions per Square Meter can be expressed as:
-            
-            $$ 
-            \text{Embodied Carbon Emissions per Square Meter (kg/m²)} = \frac{\text{Total Embodied Carbon Emissions (kg)}}{\text{Usable Floor Area (m²)}}
-            $$
-        """, unsafe_allow_html=True)
-
-        # Display the computed value
-        if embodied_carbon_emissions_per_square_meter is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Value: {embodied_carbon_emissions_per_square_meter:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        #structural efficiency: load capacity per square meter
-        # Define the variables for the new formula
-        load_capacity = 800  
-        self_weight_of_structure = 500  
-
-        # Compute the formula
-        if self_weight_of_structure != 0:  # Avoid division by zero
-            load_capacity_per_square_meter = load_capacity / self_weight_of_structure
-        else:
-            load_capacity_per_square_meter = None  # Handle division by zero case
-
-        # Structure-specific chart
-
-        st.markdown("<h2 style='text-align: center;'>Structural Efficiency: Load Capacity per Square Meter</h2>", unsafe_allow_html=True)
-        structure_data = pd.DataFrame({
-            'Load Capacity': [10],
-            'Self Weight of Structure': [1000]
-        })
-        structure_chart = px.bar(structure_data, x='Load Capacity', y='Self Weight of Structure')
-        structure_chart.update_layout(
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-        st.plotly_chart(structure_chart, use_container_width=True)
-
-        # Add major text underneath the metrics
-        st.markdown("<h2 style='text-align: center; font-size: 24px;'>Structural Flexibility: Load Capacity per Square Meter</h2>", unsafe_allow_html=True)
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the Load Capacity per Square Meter can be expressed as:
-            
-            $$ 
-            \text{Load Capacity per Square Meter (kg/m²)} = \frac{\text{Load Capacity (kg)}}{\text{Self Weight of Structure (kg)}}
-            $$
-        """, unsafe_allow_html=True)
-
-        # Display the computed value
-        if load_capacity_per_square_meter is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Value: {load_capacity_per_square_meter:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        #structural efficiency: material efficiency ratio
-        # Define the variables for the new formula
-        theoretical_minimum_material_usage = 800  
-        actual_material_usage = 500  
-
-        # Compute the formula
-        if actual_material_usage != 0:  # Avoid division by zero
-            material_efficiency_ratio = theoretical_minimum_material_usage / actual_material_usage
-        else:
-            material_efficiency_ratio = None  # Handle division by zero case
-
-        # Structure-specific chart
-
-        st.markdown("<h2 style='text-align: center;'>Structural Efficiency: Load Capacity per Square Meter</h2>", unsafe_allow_html=True)
-        structure_data = pd.DataFrame({
-            'Theoretical Minimum Material Usage': [10],
-            'Actual Material Usage': [1000]
-        })
-        structure_chart = px.bar(structure_data, x='Theoretical Minimum Material Usage', y='Actual Material Usage')
-        structure_chart.update_layout(
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-        st.plotly_chart(structure_chart, use_container_width=True)
-
-        # Add major text underneath the metrics
-        st.markdown("<h2 style='text-align: center; font-size: 24px;'>Structural Flexibility: Material Efficiency Ratio</h2>", unsafe_allow_html=True)
-
-        # Add the math formula
-        st.markdown(r"""
-            The formula for calculating the Material Efficiency Ratio can be expressed as:
-            
-            $$ 
-            \text{Material Efficiency Ratio} = \frac{\text{Theoretical Minimum Material Usage (kg)}}{\text{Actual Material Usage (kg)}}
-            $$
-        """, unsafe_allow_html=True)
-
-        # Display the computed value
-        if material_efficiency_ratio is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Calculated Value: {material_efficiency_ratio:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
+        structure_dashboard.run(metric_col1, metric_col2, selected_team)
     elif selected_team == "Industrial":
-        # Define the variables for the new formula
-        energy_generation = [800, 850, 900, 950]  # Example values over time
-        energy_demand = [500, 600, 700, 800]  # Example values over time
-
-        # Compute the energy self-sufficiency ratio for each time point
-        energy_self_sufficiency_ratios = [
-            gen / demand if demand != 0 else None
-            for gen, demand in zip(energy_generation, energy_demand)
-        ]
-
-        # Create a DataFrame for the line chart
-        time_points = ['Q1', 'Q2', 'Q3', 'Q4']  # Example time points
-        line_data = pd.DataFrame({
-            'Time': time_points,
-            'Energy Generation': energy_generation,
-            'Energy Demand': energy_demand,
-            'Energy Self-Sufficiency Ratio': energy_self_sufficiency_ratios
-        })
-
-        # Create a line chart
-        line_chart = px.line(line_data, x='Time', y=['Energy Generation', 'Energy Demand', 'Energy Self-Sufficiency Ratio'],
-                              labels={'value': 'Energy (kWh)', 'variable': 'Metrics'},
-                              title='Energy Metrics Over Time')
-        line_chart.update_layout(
-            title_x=0.5,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(line_chart, use_container_width=True)
-
-        # Display the last calculated value for energy self-sufficiency ratio
-        if energy_self_sufficiency_ratios[-1] is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Latest Energy Self-Sufficiency Ratio: {energy_self_sufficiency_ratios[-1]:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Define the variables for the new formula
-        food_production = [800, 850, 900, 950]  # Example values over time
-        food_demand = [500, 600, 700, 800]  # Example values over time
-
-        # Compute the food self-sufficiency ratio for each time point
-        food_self_sufficiency_ratios = [
-            gen / demand if demand != 0 else None
-            for gen, demand in zip(food_production, food_demand)
-        ]
-
-        # Create a DataFrame for the line chart
-        time_points = ['Q1', 'Q2', 'Q3', 'Q4']  # Example time points
-        line_data = pd.DataFrame({
-            'Time': time_points,
-            'Food Production': food_production,
-            'Food Demand': food_demand,
-            'Food Self-Sufficiency Ratio': food_self_sufficiency_ratios
-        })
-
-        # Create a line chart
-        line_chart = px.line(line_data, x='Time', y=['Food Production', 'Food Demand', 'Food Self-Sufficiency Ratio'],
-                              labels={'value': 'Food (kg)', 'variable': 'Metrics'},
-                              title='Food Metrics Over Time')
-        line_chart.update_layout(
-            title_x=0.5,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(line_chart, use_container_width=True)
-
-        # Display the last calculated value for food self-sufficiency ratio
-        if food_self_sufficiency_ratios[-1] is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Latest Food Self-Sufficiency Ratio: {food_self_sufficiency_ratios[-1]:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Define the variables for the new formula
-        recycled_water = [800, 850, 900, 950]  # Example values over time
-        wastewater_production = [500, 600, 700, 800]  # Example values over time
-
-        # Compute the recycled water ratio for each time point
-        recycled_water_ratios = [
-            gen / demand if demand != 0 else None
-            for gen, demand in zip(recycled_water, wastewater_production)
-        ]
-
-        # Create a DataFrame for the line chart
-        time_points = ['Q1', 'Q2', 'Q3', 'Q4']  # Example time points
-        line_data = pd.DataFrame({
-            'Time': time_points,
-            'Recycled Water': recycled_water,
-            'Wastewater Production': wastewater_production,
-            'Recycled Water Ratio': recycled_water_ratios
-        })
-
-        # Create a line chart
-        line_chart = px.line(line_data, x='Time', y=['Recycled Water', 'Wastewater Production', 'Recycled Water Ratio'],
-                              labels={'value': 'Recycled Water (m³)', 'variable': 'Metrics'},
-                              title='Recycled Water Metrics Over Time')
-        line_chart.update_layout(
-            title_x=0.5,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(line_chart, use_container_width=True)
-
-        # Display the last calculated value for recycled water ratio
-        if recycled_water_ratios[-1] is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Latest Recycled Water Ratio: {recycled_water_ratios[-1]:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Define the variables for the new formula
-        recycled_solid_waste = [800, 850, 900, 950]  # Example values over time
-        solid_waste_production = [500, 600, 700, 800]  # Example values over time
-
-        # Compute the solid waste utilization ratio for each time point
-        solid_waste_utilization_ratios = [
-            gen / demand if demand != 0 else None
-            for gen, demand in zip(recycled_solid_waste, solid_waste_production)
-        ]
-
-        # Create a DataFrame for the line chart
-        time_points = ['Q1', 'Q2', 'Q3', 'Q4']  # Example time points
-        line_data = pd.DataFrame({
-            'Time': time_points,
-            'Recycled Solid Waste': recycled_solid_waste,
-            'Solid Waste Production': solid_waste_production,
-            'Solid Waste Utilization Ratio': solid_waste_utilization_ratios
-        })
-
-        # Create a line chart
-        line_chart = px.line(line_data, x='Time', y=['Recycled Solid Waste', 'Solid Waste Production', 'Solid Waste Utilization Ratio'],
-                              labels={'value': 'Recycled Solid Waste (kg)', 'variable': 'Metrics'},
-                              title='Solid Waste Utilization Metrics Over Time')
-        line_chart.update_layout(
-            title_x=0.5,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(line_chart, use_container_width=True)
-
-        # Display the last calculated value for solid waste utilization ratio
-        if solid_waste_utilization_ratios[-1] is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Latest Solid Waste Utilization Ratio: {solid_waste_utilization_ratios[-1]:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
+        industrial_dashboard.run(metric_col1, metric_col2, selected_team)
 
     elif selected_team == "Facade":
-        # Define the variables for daylight in different parts of the building
-        time_of_day = ['6 AM', '9 AM', '12 PM', '3 PM', '6 PM']  # Example time points
-        daylight_amounts = {
-            'North Side': [50, 70, 90, 60, 30],  # Example daylight amounts for the North side
-            'South Side': [30, 50, 80, 100, 70],  # Example daylight amounts for the South side
-            'East Side': [60, 80, 100, 90, 40],  # Example daylight amounts for the East side
-            'West Side': [40, 60, 70, 80, 50]    # Example daylight amounts for the West side
-        }
-
-        # Total incoming daylight (example value)
-        total_incoming_daylight = 400  # Example total incoming daylight value
-
-        # Create a DataFrame for the line chart
-        daylight_data = pd.DataFrame(daylight_amounts, index=time_of_day)
-
-        # Calculate the total daylight amount used
-        total_daylight_used = daylight_data.sum().sum()  # Sum of all daylight amounts used
-
-        # Compute the ratio of daylight amount used to total incoming daylight
-        daylight_ratio = total_daylight_used / total_incoming_daylight if total_incoming_daylight != 0 else None
-
-        # Create a line chart
-        daylight_chart = px.line(daylight_data, 
-                                  title='Daylight Amount in Different Parts of the Building',
-                                  labels={'value': 'Daylight Amount (lux)', 'variable': 'Building Side'},
-                                  markers=True)
-        daylight_chart.update_layout(
-            title_x=0.35,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',
-            plot_bgcolor='rgba(0,0,0,0)',
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(daylight_chart, use_container_width=True)
-
-        # Display the daylight ratio metric
-        if daylight_ratio is not None:
-            st.markdown(f"<h3 style='text-align: center;'>Daylight Utilization Ratio: {daylight_ratio:.2f}</h3>", unsafe_allow_html=True)
-        else:
-            st.markdown("<h3 style='text-align: center;'>Error: Division by zero in calculation</h3>", unsafe_allow_html=True)
-
-        # Optionally, display a summary of the daylight amounts
-        st.markdown("<h3 style='text-align: center;'>Daylight Amounts Summary</h3>", unsafe_allow_html=True)
-        st.write(daylight_data)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-        
-        # Add the new metric for Normalized Value: Combined Metric
-        st.markdown("<h2>Primary Daylight Factor and Solar Loads Control for Residential and Work Spaces</h2>", unsafe_allow_html=True)
-
-        # Example values for the calculation
-        weight_residential = 0.5  
-        weight_work = 0.5  
-        residential_area_with_daylight = 100  
-        total_residential_area = 200  
-        work_area_with_daylight = 150  
-        total_work_area = 300  
-
-        # Calculate the normalized value
-        normalized_value = (
-            weight_residential * (residential_area_with_daylight / total_residential_area) +
-            weight_work * (work_area_with_daylight / total_work_area) * (10 / 7)
-        )
-
-        # Add the formula for the normalized value
-        st.markdown(r"""
-            The formula for calculating the Normalized Value can be expressed as:
-    
-            $$ 
-            \text{Normalized Value} = 
-            weight_{\text{residential}} \cdot \frac{\text{ResidentialAreaWithDaylight}}{\text{TotalResidentialArea}} + 
-            weight_{\text{work}} \cdot \frac{\text{WorkAreaWithDaylight}}{\text{TotalWorkArea}} \cdot \frac{10}{7}
-            $$
-    """, unsafe_allow_html=True)
-
-        # Display the normalized value
-        st.metric("Normalized Value: Combined Metric", f"{normalized_value:.2f}", help="Combined metric based on residential and work areas with daylight")
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Define the angles and their corresponding effectiveness for each panel
-        angles = [0, 15, 30, 45, 60, 75, 90]  # Example angles in degrees
-        effectiveness = [0.1, 0.5, 0.8, 1.0, 0.7, 0.4, 0.2]  # Example effectiveness values
-
-        # Create a DataFrame for the line chart
-        angle_data = pd.DataFrame({
-            'Angle (degrees)': angles,
-            'Effectiveness': effectiveness
-        })
-
-        # Compute the optimal angle (e.g., the angle with the highest effectiveness)
-        optimal_angle = angles[effectiveness.index(max(effectiveness))]
-
-        # Create a line chart
-        angle_chart = px.line(angle_data, 
-                               x='Angle (degrees)', 
-                               y='Effectiveness', 
-                               title='Effectiveness of Panel Angles',
-                               markers=True)
-        angle_chart.update_layout(
-            title_x=0.5,  # Center the title
-            paper_bgcolor='rgba(0,0,0,0)',  # Set paper background to transparent
-            plot_bgcolor='rgba(0,0,0,0)',   # Set plot background to transparent
-            font_family="Roboto Mono",
-            font_color="#2c3e50"
-        )
-
-        # Display the line chart
-        st.plotly_chart(angle_chart, use_container_width=True)
-
-        # Display the optimal angle metric
-        st.markdown(f"<h3 style='text-align: center;'>Optimal Angle for Panel Movement: {optimal_angle}°</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
-        # Example values for energy generation and energy required
-        energy_generation = 1500  
-        energy_required_by_industrial_team = 1000  
-
-        # Calculate the ratio
-        ratio = energy_generation / energy_required_by_industrial_team
-
-        # Create a DataFrame for the bar chart
-        data = {
-            'Metric': ['Energy Generation', 'Energy Required by Industrial Team'],
-            'Value': [energy_generation, energy_required_by_industrial_team]
-        }
-
-        df = pd.DataFrame(data)
-
-        # Create the bar chart
-        fig = px.bar(df, x='Metric', y='Value',
-             labels={'Value': 'Energy (kWh)', 'Metric': 'Metrics'},
-             color='Metric')
-
-        # Add the ratio as a text annotation
-        fig.add_annotation(
-            x='Energy Generation',
-            y=ratio,
-            text=f'Ratio: {ratio:.2f}',
-            showarrow=True,
-            arrowhead=2,
-            ax=0,
-            ay=-40
-            )
-
-        # Display the chart in Streamlit
-        st.plotly_chart(fig, use_container_width=True)
-
-        st.markdown(f"<h3 style='text-align: center;'>Energy Generation Ratio: {ratio:.2f}</h3>", unsafe_allow_html=True)
-
-        st.markdown("")
-        st.markdown("")
-
-        st.markdown("<hr>", unsafe_allow_html=True)  # Add a line after the metric
-
+        facade_dashboard.run(metric_col1, metric_col2, selected_team)
 print(f'selected_model: {selected_model}\n')
 print(f'selected_version: {selected_version}\n')
